+++ mayan/apps/file_metadata/handlers.py
@@ -1,13 +1,12 @@
 from django.apps import apps
 
-from .settings import setting_auto_process
-
 
 def handler_initialize_new_document_type_file_metadata_driver_configuration(
     sender, instance, **kwargs
 ):
     DocumentTypeDriverConfiguration = apps.get_model(
-        app_label='file_metadata', model_name='DocumentTypeDriverConfiguration'
+        app_label='file_metadata',
+        model_name='DocumentTypeDriverConfiguration'
     )
     StoredDriver = apps.get_model(
         app_label='file_metadata', model_name='StoredDriver'
@@ -15,8 +14,11 @@ def handler_initialize_new_document_type_file_metadata_driver_configuration(
 
     if kwargs['created']:
         for stored_driver in StoredDriver.objects.all():
+            driver_class = stored_driver.driver_class
+            enabled = driver_class.get_enabled_value()
+
             DocumentTypeDriverConfiguration.objects.create(
-                document_type=instance, enabled=setting_auto_process.value,
+                document_type=instance, enabled=enabled,
                 stored_driver=stored_driver
             )